+++ built_value_generator/lib/built_value_generator.dart
@@ -131,6 +131,7 @@ class BuiltValueGenerator extends Generator {
     final result = <FieldElement>[];
     for (final field in classElement.fields) {
       if (!field.isStatic &&
+          field.getter != null &&
           (field.getter.isAbstract || field.getter.isSynthetic)) result
           .add(field);
     }