+++ example/lib/value.g.dart
@@ -1,5 +1,4 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// 2016-02-23T10:04:23.739Z
 
 part of value;